web/backend/requirements.txt
django
django-cors-headers
djangorestframework
maude